requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
numpy==1.26.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pycparser==2.21
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyVirtualDisplay==3.0
requests==2.31.0
selenium==4.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
urllib3-secure-extra==0.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
